Support for organizations

+++ app.py
@@ -1,20 +1,23 @@
-import os, requests, subprocess, zipfile
-from flask import Flask, Response, request, send_file, send_from_directory
+import os, subprocess, shutil
+from flask import Flask, Response, request, send_from_directory
 from dotenv import load_dotenv
 
 load_dotenv()
 app = Flask(__name__)
 
-
 @app.route('/')
 def ping():
-    status_code = Response(status=200)
-    return status_code
+    return Response(status=200)
 
+# Returns the processing status based on the mesh generation's log file
 @app.route('/process-status/', methods=['GET'])
 def get_pre_processor_status():
-    context_name = request.args.get('context_name')
-    log_file = f"logs/{context_name}_process_log.txt"
+    organizationCode = request.args.get('organizationCode', False)
+    contextCode = request.args.get('contextCode', False)
+    if not organizationCode or not contextCode:
+        return Response(status=400)
+
+    log_file = f'logs/{organizationCode}/{contextCode}/mesh_log.txt'
     if not os.path.isfile(log_file):
         return Response(status=404)
 
@@ -24,30 +27,41 @@ def get_pre_processor_status():
     
     return msg
 
+# Initiates a mesh generation process
 @app.route('/process/', methods=['POST'])
 def run_pre_processor():
-    # requester_ip = request.remote_addr
     requester_ip = os.environ['RIVERCURE_URL']
-    context_name = request.args.get('context_name')
-    destination_folder = f'{context_name}_simulation/gis/'
-    print(context_name)
-    log_file = f"logs/{context_name}_process_log.txt"
+    organizationCode = request.args.get('organizationCode', False)
+    contextCode = request.args.get('contextCode', False)
+    if not organizationCode or not contextCode:
+        return Response(status=400)
+
+    destination_folder = f'{organizationCode}/{contextCode}_simulation/gis'
+    print(f'Mesh generate request for context {contextCode} of organization {organizationCode}')
+
+    log_file = f'logs/{organizationCode}/{contextCode}/mesh_log.txt'
+
+    # If organization's/context's log folder doesnt exist, create
+    if not os.path.exists(f'logs/{organizationCode}/{contextCode}'):
+        os.makedirs(f'logs/{organizationCode}/{contextCode}')
+
+    # Delete log file for this context if it already exists
     if os.path.exists(log_file):
         os.remove(log_file)
 
     try:
-        os.system(f'cp -r simulation {context_name}_simulation')
+        shutil.copytree('simulation', f'{organizationCode}/{contextCode}_simulation', dirs_exist_ok=True)
 
         for key in request.files:
             if key == 'dtm.tif' or key == 'frictionCoef.tif':
-                request.files[key].save(f'{destination_folder}rasters/{key}')
+                request.files[key].save(f'{destination_folder}/rasters/{key}')
             else:
-                request.files[key].save(f'{destination_folder}context_files/{key}')
+                request.files[key].save(f'{destination_folder}/context_files/{key}')
 
-        log_f = open(log_file, "w")
-        subprocess.Popen(f'''(cd {context_name}_simulation/gis && ./mesh && cd ../.. \
-                && cp {context_name}_simulation/mesh/vtk/meshQuality.vtk paraview_visualizer/data/{context_name}_mesh.vtk \
-                && curl {requester_ip}/contexts/mesh-status/{context_name}?status=True &)''', 
+        log_f = open(log_file, 'w')
+        subprocess.Popen(f'''(cd {organizationCode}/{contextCode}_simulation/gis && ./mesh && cd ../../.. \
+                && cp {organizationCode}/{contextCode}_simulation/mesh/vtk/meshQuality.vtk paraview_visualizer/data/{organizationCode}_{contextCode}_mesh.vtk \
+                && curl {requester_ip}/contexts/mesh-status/{contextCode}/change?organization={organizationCode}\&status=True &)''', 
                 stdout=log_f, stderr=log_f, shell=True)
         
     except Exception as e:
@@ -58,34 +72,61 @@ def run_pre_processor():
 
 @app.route('/simulate-status/', methods=['GET'])
 def get_simulation_status():
-    event_id = request.args.get('event_id')
-    context_name = request.args.get('context_name')
-    log_file = f"logs/{context_name}_simulate_{event_id}_log.txt"
+    organizationCode = request.args.get('organizationCode', False)
+    contextCode = request.args.get('contextCode', False)
+    eventName = request.args.get('eventName', False)
+    if not organizationCode or not contextCode or not eventName:
+        return Response(status=400)
+    
+    # Remove whitespaces from event name
+    eventName = eventName.replace(' ', '_')
+
+    log_file = f'logs/{organizationCode}/{contextCode}/{eventName}_simulation_log.txt'
     if not os.path.isfile(log_file):
         return Response(status=404)
 
+    cleanup = False
     msg = []
-    with open(log_file, "r") as f_log:
-        msg = f_log.readlines()[-100:]
+    with open(log_file, 'r') as f_log:
+        lines = f_log.readlines()
+        msg = lines[-100:]
+        numberOfLines = len(lines)
+        if numberOfLines > 10000:
+            cleanup = True
     msg = "".join(msg)
+
+    # Prevent file from oversizing
+    if cleanup:
+        open(log_file, 'w').close()
     
     return msg
 
 @app.route('/simulate/', methods=['POST'])
 def simulate():
     requester_ip = os.environ['RIVERCURE_URL']
-    context_name = request.args.get('context_name')
-    event_id = request.args.get('event_id')
-    print(f'Simulation request for event {event_id} of context {context_name}')
+    organizationCode = request.args.get('organizationCode', False)
+    contextCode = request.args.get('contextCode', False)
+    eventName = request.args.get('eventName', False)
+    eventId = request.args.get('eventId', False)
+    if not organizationCode or not contextCode or not eventName or not eventId:
+        return Response(status=400)
+        
+    print(f'Simulation request for event {eventName}, of context {contextCode}, of organization {organizationCode}')
+
+    # Remove whitespaces from event name
+    eventName = eventName.replace(' ', '_')
     
-    frequency_destination_folder = f'{context_name}_simulation/output/output.cnt'
-    sensor_data_destination_folder = f'{context_name}_simulation/boundary/gauges'
-    time_destination_folder = f'{context_name}_simulation/control/time.cnt'
-    boundary_destination_folder = f'{context_name}_simulation/boundary/boundary.cnt'
+    frequency_destination_folder   = f'{organizationCode}/{contextCode}_simulation/output/output.cnt'
+    sensor_data_destination_folder = f'{organizationCode}/{contextCode}_simulation/boundary/gauges'
+    time_destination_folder        = f'{organizationCode}/{contextCode}_simulation/control/time.cnt'
+    boundary_destination_folder    = f'{organizationCode}/{contextCode}_simulation/boundary/boundary.cnt'
     
-    log_file = f"logs/{context_name}_simulate_{event_id}_log.txt"
+    log_file = f'logs/{organizationCode}/{contextCode}/{eventName}_simulation_log.txt'
     if os.path.exists(log_file):
         os.remove(log_file)
+    
+    # Creates folders if they dont exist
+    os.makedirs(f'logs/{organizationCode}/{contextCode}', exist_ok=True)
     # bnd are duplicated might be necessary to remove them
     try:
         if request.method == 'POST':
@@ -99,10 +140,10 @@ def simulate():
                 else:
                     request.files[key].save(f'{sensor_data_destination_folder}/{key}')
 
-        log_f = open(log_file, "w")
+        log_f = open(log_file, 'w')
         # curl {requester_ip}/contexts/simulation/results/handle/{event_id} used to be here but now is not needed. Only the VTK in output/maxima/ is.
-        subprocess.Popen(f'''(cd {context_name}_simulation && ./solver2D \
-                            && curl {requester_ip}/contexts/event-status/{event_id}?status=True &)''', 
+        subprocess.Popen(f'''(cd {organizationCode}/{contextCode}_simulation && ./solver2D \
+                            && curl {requester_ip}/contexts/event-status/{eventId}/change?status=True &)''', 
                             stdout=log_f, stderr=log_f, shell=True)
     except Exception as e:
         print(f'Failed simulation!\nException{e}')
@@ -112,14 +153,22 @@ def simulate():
 
 @app.route('/pre-processing/results/')
 def preprocessing_results():
-    context_name = request.args.get('context_name')
-    path = os.path.join(app.root_path, f'{context_name}_simulation/mesh/vtk/')
+    organizationCode = request.args.get('organizationCode', False)
+    contextCode = request.args.get('contextCode', False)
+    if not organizationCode or not contextCode:
+        return Response(status=400)
+
+    path = os.path.join(app.root_path, f'{organizationCode}/{contextCode}_simulation/mesh/vtk/')
     return send_from_directory(path, filename='meshQuality.vtk')
 
 @app.route('/simulation/results/')
 def simulation_results():
-    context_name = request.args.get('context_name')
-    path = os.path.join(app.root_path, f'{context_name}_simulation/output/maxima/')
+    organizationCode = request.args.get('organizationCode', False)
+    contextCode = request.args.get('contextCode', False)
+    if not organizationCode or not contextCode:
+        return Response(status=400)
+
+    path = os.path.join(app.root_path, f'{organizationCode}/{contextCode}_simulation/output/maxima/')
     # The name can vary, but always starts with 'maxi'. Assume only one file, the desired one, is in this folder
     try:
         file_name = os.listdir(path)[0]
